Finalizing inventory uncertainty calculation for raster and vector datasets
This commit features first complete functionality for Gaussian error
propagation of inventories. Inventory imports as arrays, dictionaries,
vector and raster datasets are supported. The propagation can be
executed with our without covariances (in case of spatial independence).
The spatial autocorrelation can be tested with the Moran's I statistic
beforehand of the error propagation. The calculation of covariance
matrices for spatial datasets via empirical and theroretical (spherical)
semivariograms is implemented and tested.